Add files via upload
The Hospital Record Management System is a Java-based application developed to facilitate efficient management of patient records within a hospital setting. Leveraging JDBC (Java Database Connectivity) for database interaction and AWT (Abstract Window Toolkit) for creating the user interface, this system provides functionalities for connecting to a database, retrieving patient data, and displaying it in an organized manner. 

The system connects seamlessly to a database management system (DBMS) to store and retrieve patient records. JDBC provides the necessary framework to establish this connection.